add useParams in ProductDetail, fix avatar in Review
add useParams in ProductDetail, fix avatar in Review
