+++ src/components/ProductDetail/ProductDetail.module.scss
@@ -211,11 +211,23 @@
   margin-bottom: 8px;
 }
 
-.avatar {
+.avatarWrapper {
   width: 40px;
   height: 40px;
   margin-right: 12px;
-  margin-left: 16px;
+  border-radius: 50%;
+  padding: 0;
+  overflow: hidden;
+  fill: transparent;
+  background-color: transparent;
+}
+
+.avatar {
+  width: 40px;
+  height: 40px;
+  border-radius: 50%;
+  margin-right: auto;
+  margin-left: auto;
 }
 
 .customAvatarStyle {
